requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
cffi==1.16.0
channels==4.2.0
channels-redis==4.2.0
click==8.1.7
configparser==3.5.0
constantly==23.10.4
crispy-bootstrap4==2024.1
cryptography==42.0.8
daphne==4.1.2
Django==5.0.6
django-allauth==0.63.3
django-bootstrap4==24.3
django-core==1.4.1
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-environ==0.11.2
django-extensions==3.2.3
django-filebrowser==4.0.3
django-filter==24.2
django-grappelli==3.0.9
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
django-stub==0.1
django-stubs==5.1.3
django-stubs-ext==5.1.3
django-tables2==2.7.0
django-taggit==5.0.1
django-tinymce==4.0.0
djangorestframework==3.15.2
djangosecure==0.0.11
future==0.16.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
msgpack==1.0.8
pillow==10.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.6.1
pyOpenSSL==24.1.0
pytz==2024.1
python-dotenv==1.0.1
redis==5.0.6
rq==1.16.2
service-identity==24.1.0
setuptools==70.1.0
six==1.10.0
soupsieve==2.5
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
types-PyYAML==6.0.12.20250326
typing_extensions==4.12.2
whitenoise==6.6.0
zope.interface==6.4.post2